portfolio/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
Jinja2==3.1.2
python-multipart==0.0.6
httpx==0.27.0
pytest>=8.2.0,<10
pytest-asyncio==1.3.0
Pillow==10.1.0
